refactor(blorktools): rig anmiation to manager
