+++ packages/blorktools/src/asset_debugger/core/rig/rig-manager.js
@@ -1,5 +1,7 @@
 import * as THREE from 'three';
-import { getIsDragging } from '../drag-util';
+import { checkHandleHover, getIsDragging } from '../drag-util';
+import { getState } from '../state';
+import { restoreLockedBoneRotations } from '../../ui/rig-panel';
 
 // These variables need to be exported so they're available to both modules
 export let bones = [];
@@ -23,6 +25,59 @@ export const rigOptions = {
     hoverColor: 0x00FF00,  // Green - for control handles hover state
     activeColor: 0x0000FF  // Blue - for control handles active/dragging state
 };
+
+/**
+ * Update animation for rig visuals
+ */
+export function updateRigAnimation() {
+    // Only update rig visuals if Display Rig is enabled
+    if (!rigOptions.displayRig) {
+        // Even when displayRig is off, we should ensure handles are not visible
+        if (furthestBoneHandle) {
+            furthestBoneHandle.visible = false;
+        }
+        if (boneVisualsGroup) {
+            boneVisualsGroup.visible = false;
+        }
+        return;
+    }
+    
+    // Update bone visuals
+    if (boneVisualsGroup) {
+        boneVisualsGroup.visible = true;
+        boneVisualsGroup.children.forEach(boneGroup => {
+            if (boneGroup.userData.updatePosition) {
+                boneGroup.userData.updatePosition();
+            }
+        });
+    }
+    
+    // Update furthest bone handle
+    if (furthestBoneHandle) {
+        furthestBoneHandle.visible = true;
+        if (furthestBoneHandle.userData.updatePosition && !getIsDragging()) {
+            // Only update handle position when not dragging
+            furthestBoneHandle.userData.updatePosition();
+        }
+    }
+    
+    // Update joint labels
+    const state = getState();
+    const labelGroup = state.scene ? state.scene.getObjectByName("JointLabels") : null;
+    if (labelGroup) {
+        labelGroup.children.forEach(label => {
+            if (label.userData && label.userData.updatePosition) {
+                label.userData.updatePosition();
+            }
+        });
+    }
+    
+    // Apply locked rotations to bones
+    restoreLockedBoneRotations();
+    
+    // Check handle hover on each frame
+    checkHandleHover();
+}
 
 /**
  * Clear existing rig visualization from the scene
